+++ reactsite.Service/Interfaces/IDailyTasksService.cs
@@ -12,6 +12,6 @@ namespace reactsite.Service.Interfaces
     public interface IDailyTasksService
     {
         /*Task<BaseResponse<bool>> CreateProfile(ProfileViewModel p);*/
-        Task<BaseResponse<DailyTasksViewModel>> GetDailyTask(int userid);
+        Task<BaseResponse<DailyTasksViewModel>> GetDailyTask(long userid);
     }
 }
